Refactor: 로그인, 회원가입 디렉터리 구조
