+++ src/main/java/capstone/recipable/domain/auth/config/MailConfig.java
@@ -1,4 +1,4 @@
-package capstone.recipable.domain.email.config;
+package capstone.recipable.domain.auth.config;
 
 
 import org.springframework.beans.factory.annotation.Value;
