+++ src/main/java/capstone/recipable/domain/auth/dto/request/RegisterRequest.java
@@ -1,4 +1,4 @@
-package capstone.recipable.domain.register.dto.request;
+package capstone.recipable.domain.auth.dto.request;
 
 import capstone.recipable.domain.user.entity.User;
 import lombok.AllArgsConstructor;
